rework: Organize pages
